backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pillow>=10.2.0
easyocr==1.7.0
transformers>=4.37.0
torch>=2.2.0
sentencepiece>=0.1.99
sqlalchemy==2.0.23
pydantic>=2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
numpy>=1.24.3
pandas>=2.1.3
certifi>=2023.0.0
google-generativeai>=0.3.2